Logic improvement
